RDKTT-2707 : Add offset in Performance scripts
Reason for change: Add offset for threshold values in Performance scripts. Since these scripts are measuring the time taken for each operations in milliseconds, variations to an extent is acceptable. So we need to add an offset of configurable value.
Test Procedure: Check if test success with TDK
Risks: None
Signed-off-by:Abin T S<abin.ts@tataelxsi.co.in>

Change-Id: Ia09388cea6d3f7e037a446d66e246be77cbb76aa